Auto-vivify nonexistent metrics
This is a handy feature to have, but Folsom's support doesn't work quite right.
In particular, there's no way to teach it to auto-vivify a slide_uniform
histogram.

Furthermore, we should be catching on metric creates to ensure that a creation
attempt made while, e.g., Folsom is being restarted doesn't crash processes.